services/cost-analyzer/requirements.txt
nats-py==2.7.2
redis==5.0.3
prometheus-client==0.20.0